spending-insights-budgeting/pyproject.toml
[project]
name = "spending-insights-budgeting"
version = "0.1.0"
description = "RAG-powered spending insights with DuckDB, LLM, Guardrails, MCP, and LangChain"
authors = [
    {name = "Rahul Tomar", email = "rahul@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.20",
    "openai>=1.3.0",
    "faiss-cpu>=1.7.4",
    "duckdb>=0.9.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "watchdog>=6.0.0",
    "psutil>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
